# Improve DAC timing plausibility and add a CPU timer fallback