Here it is. 
It works surprisingly well with the pulse ALSA device. Even if we have a 47 ms cycle for a 43 ms buffer and two callbacks without a gap if the buffer is consumed.