@rryan 2016-04-16
I enabled very verbose debugging in portaudio. Here's a normal run where the plausibility check passes:
```
Debug [Main]: Loading SoundSource plugins "/Users/rjryan/Code/mixxx/osx64_build/plugins"
Debug [Main]: SoundSource providers for file extension "aif"
Debug [Main]:   1 : "libsndfile"
Debug [Main]: SoundSource providers for file extension "aiff"
Debug [Main]:   1 : "libsndfile"
Debug [Main]: SoundSource providers for file extension "flac"
Debug [Main]:   2 : "Xiph.org libFLAC"
Debug [Main]:   1 : "libsndfile"
Debug [Main]: SoundSource providers for file extension "it"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "m4a"
Debug [Main]:   2 : "Apple Core Audio"
Debug [Main]: SoundSource providers for file extension "med"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "mod"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "mp2"
Debug [Main]:   2 : "Apple Core Audio"
Debug [Main]: SoundSource providers for file extension "mp3"
Debug [Main]:   2 : "Apple Core Audio"
Debug [Main]: SoundSource providers for file extension "ogg"
Debug [Main]:   2 : "Xiph.org OggVorbis"
Debug [Main]: SoundSource providers for file extension "okt"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "opus"
Debug [Main]:   2 : "Xiph.org libopusfile"
Debug [Main]: SoundSource providers for file extension "s3m"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "stm"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: SoundSource providers for file extension "wav"
Debug [Main]:   1 : "libsndfile"
Debug [Main]: SoundSource providers for file extension "xm"
Debug [Main]:   2 : "MODPlug"
Debug [Main]: qt_mac_get_scalefactor 2
Debug [Main]: "Mixxx" "2.1.0-alpha-pre" "(git clockref r5618; built on: Apr 16 2016 @ 00:29:01; flags: asan=0 autodjcrates=1 battery=0 buildtime=1 bulk=0 color=0 coreaudio=1 faad=1 ffmpeg=0 hid=1 hss1394=0 ipod=0 localecompare=0 macappstore=0 mad=0 mediafoundation=0 modplug=1 opengles=0 optimize=native opus=1 perftools=0 perftools_profiler=0 profiling=0 qdebug=1 qt_sqlite_plugin=1 shoutcast=1 test=1 vamp=1 verbose=0 vinylcontrol=1 wv=1)" is starting...
Debug [Main]: Library versions:
Debug [Main]: Qt: 4.8.7
Debug [Main]: libshout: 2.3.1
Debug [Main]: PortAudio: 1246464 PortAudio V19.5.0-devel, revision unknown
Debug [Main]: RubberBand: 1.8.1
Debug [Main]: SoundTouch: 1.9.2
Debug [Main]: TagLib: 1.9.1
Debug [Main]: ChromaPrint: 1.2.0
Debug [Main]: Vorbis: Xiph.Org libVorbis 1.3.5
Debug [Main]: FLAC: 1.3.1
Debug [Main]: QDesktopServices::storageLocation(HomeLocation): "/Users/rjryan"
Debug [Main]: QDesktopServices::storageLocation(DataLocation): "/Users/rjryan/Library/Application Support/Mixxx"
Debug [Main]: QCoreApplication::applicationDirPath() "/Users/rjryan/Code/mixxx"
Debug [Main]: Loading resources from  "/Users/rjryan/Code/mixxx/res/"
Warning [Main]: Configuration file is at version "2.0.0" instead of the current 2.1.0-alpha-pre
Debug [Main]: Loading resources from  "/Users/rjryan/Code/mixxx/res/"
Warning [Main]: QFileInfo::absolutePath: Constructed with empty filename
Debug [Main]: Loading resources from  "/Users/rjryan/Code/mixxx/res/"
Debug [Main]: Loading translations for locale "en_US" from translations folder "/Users/rjryan/Code/mixxx/res/translations/" : fail
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Samplers]" , "show_samplers" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Microphone]" , "show_microphone" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[VinylControl]" , "show_vinylcontrol" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[PreviewDeck]" , "show_previewdeck" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[EffectRack1]" , "show" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Library]" , "show_coverart" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Master]" , "maximize_library" )
Debug [Main]: Loading resources from  "/Users/rjryan/Code/mixxx/res/"
Debug [Main]: Compressor attack per frame:  0.000408163 decay per frame:  4.08163e-05
before paHostApiInitializers[0].
||PaMacCore (vv)|| PaMacCore_Initialize(): hostApiIndex=0
||PaMacCore (vv)|| gatherDeviceInfo()
||PaMacCore (v )|| Found 2 device(s).
Device 0	: 43
Device 1	: 50
||PaMacCore (v )|| Default in : 43
||PaMacCore (v )|| Default out: 50
||PaMacCore (vv)|| InitializeDeviceInfo(): macCoreDeviceId=43
||PaMacCore (vv)|| GetChannelInfo()
||PaMacCore (vv)|| GetChannelInfo()
||PaMacCore (vv)|| InitializeDeviceInfo(): macCoreDeviceId=50
||PaMacCore (vv)|| GetChannelInfo()
||PaMacCore (vv)|| GetChannelInfo()
after paHostApiInitializers[0].
Debug [Main]: WARNING: AudioInput already registered!
Debug [Main]: WARNING: AudioInput already registered!
Debug [Main]: WARNING: AudioInput already registered!
Debug [Main]: WARNING: AudioInput already registered!
Debug [Main]: Available QtSQL drivers: ("QSQLITE", "QPSQL7", "QPSQL")
Debug [Main]: DB status: "/Users/rjryan/Library/Application Support/Mixxx/mixxxdb.sqlite" = true
Debug [Main]: SchemaManager::upgradeToSchemaVersion already at version 26
Debug [Main]: TrackDAO::initialize QThread(0x7fbba8c6f7c0, name = "Main") "qt_sql_default_connection"
Debug [Main]: CrateDAO::initialize()
Debug [Main]: CueDAO::initialize QThread(0x7fbba8c6f7c0, name = "Main") "qt_sql_default_connection"
Debug [Main]: DirectoryDAO::initialize QThread(0x7fbba8c6f7c0, name = "Main") "qt_sql_default_connection"
Debug [Main]: LibraryHashDAO::initialize QThread(0x7fbba8c6f7c0, name = "Main") "qt_sql_default_connection"
Debug [Main]: Starting LibraryScanner thread.
Debug [LibraryScanner 1]: LibraryHashDAO::initialize LibraryScanner(0x7fbbaeb23b48, name = "LibraryScanner 1") "LIBRARY_SCANNER"
Debug [LibraryScanner 1]: CueDAO::initialize LibraryScanner(0x7fbbaeb23b48, name = "LibraryScanner 1") "LIBRARY_SCANNER"
Debug [LibraryScanner 1]: TrackDAO::initialize LibraryScanner(0x7fbbaeb23b48, name = "LibraryScanner 1") "LIBRARY_SCANNER"
Debug [LibraryScanner 1]: DirectoryDAO::initialize LibraryScanner(0x7fbbaeb23b48, name = "LibraryScanner 1") "LIBRARY_SCANNER"
Debug [LibraryScanner 1]: LibraryScanner event loop starting.
Debug [Main]: Appending Quick Link:  "Music" --- "/Users/rjryan/Music/"
Debug [Main]: Appending Quick Link:  "Downloads" --- "/Users/rjryan/Downloads/"
Debug [Main]: Appending Quick Link:  "DownloadsOld" --- "/Users/rjryan/Downloads/DownloadsOld/"
Debug [Main]: Committing transaction on "qt_sql_default_connection" result: true
Debug [Main]: Traktor Library Location=[ "/Users/rjryan/collection.nml" ]
Debug [Main]: Checking for access to user's library directories:
Debug [Main]: Checking for access to "/Users/rjryan/Drive/Music" : true
Debug [Main]: Checking for access to "/Users/rjryan/Movies" : true
Debug [Main]: AnalyzerWaveform::AnalyzerWaveform()
Debug [Main]: Setting VAMP_PATH to:  "/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins"
Debug [Main]: Creating ControllerManager
Debug [Main]: Extension ".bulk.xml" total 1 presets
Debug [Main]: Extension ".hid.xml" total 15 presets
Debug [Main]: Extension ".midi.xml" total 96 presets
Debug [Main]: Setting VAMP_PATH to:  "/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins:/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins"
Debug [Main]: VampPluginLoader::listPlugins() returned 4 plugins
Debug [Main]: Plugin output displayname: "mixxxbpmdetection:0" "SoundTouch BPM Detector (Legacy)"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:0" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:1" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:2" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:3" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-keydetector:0" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:1" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:2" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:3" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-tempotracker:0" "Tempo and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-tempotracker:1" "Tempo and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-tempotracker:2" "Tempo and Beat Tracker"
Debug [Main]: Setting VAMP_PATH to:  "/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins:/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins:/Users/rjryan/Code/mixxx/osx64_build/vamp-plugins"
Debug [Main]: VampPluginLoader::listPlugins() returned 4 plugins
Debug [Main]: Plugin output displayname: "mixxxbpmdetection:0" "SoundTouch BPM Detector (Legacy)"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:0" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:1" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:2" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-barbeattracker:3" "Bar and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-keydetector:0" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:1" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:2" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-keydetector:3" "Key Detector"
Debug [Main]: Plugin output displayname: "qm-tempotracker:0" "Tempo and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-tempotracker:1" "Tempo and Beat Tracker"
Debug [Main]: Plugin output displayname: "qm-tempotracker:2" "Tempo and Beat Tracker"
Debug [Main]: DlgPrefKey::loadSettings
Debug [Main]: Key plugin ID: "qm-keydetector:2"
Debug [Main]: ControllerManager::getControllerList
Debug [Main]: Set root GL Context widget valid: QGLWidget(0x7fbbae8f31c0) true
Debug [Main]: Created root GL Context valid: 0x7fbbae8f31f0 true
Debug [Main]: Root GL Context format:
Debug [Main]: Double Buffering: true
Debug [Main]: Swap interval: 1
Debug [Main]: Depth buffer: true
Debug [Main]: Direct rendering: true
Debug [Main]: Has overlay: false
Debug [Main]: RGBA: true
Debug [Main]: Sample buffers: false
Debug [Main]: Samples: -1
Debug [Main]: Stencil buffers: true
Debug [Main]: Stereo: false
Debug [Main]: LegacySkinParser loading skin: "/Users/rjryan/Code/mixxx/res/skins/TestKitchen"
Debug [Main]: Skin is a >=1.12.0 style skin.
Warning [Main]: src/skin/legacyskinparser.cpp:570 SKIN ERROR at /Users/rjryan/Code/mixxx/res/skins/TestKitchen/skin.xml:20 <QuickWidget>: Invalid node name in skin: "QuickWidget"
Debug [Main]: MixxxLibraryFeature::activate()
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Samplers]" , "show_samplers" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Microphone]" , "show_microphone" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[VinylControl]" , "show_vinylcontrol" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[PreviewDeck]" , "show_previewdeck" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[EffectRack1]" , "show" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Library]" , "show_coverart" )
Warning [Main]: ControlDoublePrivate::getControl returning NULL for ( "[Master]" , "maximize_library" )
Debug [Controller]: ControllerManager: Setting up devices
Debug [Main]: SoundManager::setupDevices()
Debug [Controller]: Scanning PortMIDI devices:
Debug [LibraryScanner 1]: LibraryScanner::slotStartScan
Debug [Controller]: Scanning HID devices:
Debug [Main]: SoundDevicePortAudio::open() "1, Built-in Output"
Debug [Main]: framesPerBuffer: 64
Debug [Main]: Requested sample rate:  44100 Hz, latency: 1.45125 ms
Debug [Main]: Output channels: 2 | Input channels: 0
Debug [Main]: Opening stream with id 1
||PaMacCore (vv)|| OpenStream(): in chan=-1, in fmt=-1, out chan=2, out fmt=1 SR=44100, FPB=64
||PaMacCore (v )|| Opening Stream.
||PaMacCore (v )|| After querying hardware, setting block size to 64.
||PaMacCore (vv)|| OpenAndSetupOneAudioUnit(): in chan=-1, in fmt=-1, out chan=2, out fmt=1, requestedFramesPerBuffer=64
Debug [LibraryScanner 1]: Recursively scanning library.
2016-04-16 00:44:39.881 mixxx[88769:12210066] 00:44:39.881 WARNING:  140: This application, or a library it uses, is using the deprecated Carbon Component Manager for hosting Audio Units. Support for this will be removed in a future release. Also, this makes the host incompatible with version 3 audio units. Please transition to the API's in AudioComponent.h.
||PaMacCore (v )|| Opened device 50 for output.
||PaMacCore (vv)|| OpenAndSetupOneAudioUnit(): in chan=-1, in fmt=-1, out chan=-1, out fmt=-1, requestedFramesPerBuffer=64
Debug [Controller]: Found "Apple Inc." "Apple Internal Keyboard / Trackpad" "r549" S/N "" "Usage Page 1 Usage 2"
Debug [Controller]: Found "Apple Inc." "Apple Internal Keyboard / Trackpad" "r549" S/N "" "Usage Page 65280 Usage 1"
Debug [Controller]: Found "Apple Inc." "Apple Internal Keyboard / Trackpad" "r549" S/N "" "Usage Page 1 Usage 6"
Debug [Controller]: Found "Apple Inc." "Apple Internal Keyboard / Trackpad" "r549" S/N "" "Usage Page 1 Usage 2"
Debug [Controller]: ControllerManager::getControllerList
Debug [Controller]: Searching for controller preset "Apple_Internal_Keyboard___Trackpad_.hid.xml" in paths: "/Users/rjryan/Library/Application Support/Mixxx/controllers/,/Users/rjryan/Code/mixxx/res/controllers/"
Debug [Controller]: Loading controller preset from "/Users/rjryan/Library/Application Support/Mixxx/controllers/Apple_Internal_Keyboard___Trackpad_.hid.xml"
Debug [Main]: Opened PortAudio stream successfully... starting
||PaMacCore (vv)|| IsStreamStopped()
||PaMacCore (vv)|| StartStream()
||PaMacCore (v )|| Starting stream.
Debug [Controller]: Searching for controller preset "Apple_Internal_Keyboard___Trackpad_--2.hid.xml" in paths: "/Users/rjryan/Library/Application Support/Mixxx/controllers/,/Users/rjryan/Code/mixxx/res/controllers/"
Debug [Main]: PortAudio: Started stream successfully
Debug [Main]:    Actual sample rate:  44100 Hz, latency: 13.5828 ms
Debug [Main]: SoundDeviceNetwork::open() "Network stream"
Debug [Controller]: Loading controller preset from "/Users/rjryan/Library/Application Support/Mixxx/controllers/Apple_Internal_Keyboard___Trackpad_--2.hid.xml"
Debug [Main]: framesPerBuffer: 64
Debug [Main]: Requested sample rate:  44100 Hz, latency: 1.45125 ms
Debug [Main]: Using "Built-in Output" as output sound device clock reference
Debug [Main]: 2 output sound devices opened
Debug [Main]: 0 input  sound devices opened
Debug [Controller]: Searching for controller preset "Apple_Internal_Keyboard___Trackpad_--3.hid.xml" in paths: "/Users/rjryan/Library/Application Support/Mixxx/controllers/,/Users/rjryan/Code/mixxx/res/controllers/"
Debug [Controller]: Loading controller preset from "/Users/rjryan/Library/Application Support/Mixxx/controllers/Apple_Internal_Keyboard___Trackpad_--3.hid.xml"
Debug [Main]: Displaying mixxx
Debug [Main]: Running Mixxx
Debug [Main]: ControllerManager::getControllerList
Debug [Controller]: Searching for controller preset "Apple_Internal_Keyboard___Trackpad_--4.hid.xml" in paths: "/Users/rjryan/Library/Application Support/Mixxx/controllers/,/Users/rjryan/Code/mixxx/res/controllers/"
Debug [Controller]: Loading controller preset from "/Users/rjryan/Library/Application Support/Mixxx/controllers/Apple_Internal_Keyboard___Trackpad_--4.hid.xml"
Debug [Controller]: Controller polling stopped.
Debug [LibraryScanner 1]: LibraryScanner::slotFinishHashedScan
Debug [LibraryScanner 1]: LibraryScanner::slotFinishUnhashedScan
Debug [LibraryScanner 1]: Recursive scanning finished cleanly.
Debug [LibraryScanner 1]: Committing transaction on "LIBRARY_SCANNER" result: true
Debug [LibraryScanner 1]: Marking tracks in changed directories as verified
Debug [LibraryScanner 1]: Marking unchanged directories and tracks as verified
Debug [LibraryScanner 1]: Checking remaining unverified tracks.
Debug [LibraryScanner 1]: Marking unverified tracks as deleted.
Debug [LibraryScanner 1]: Marking unverified directories as deleted.
Debug [LibraryScanner 1]: Detecting moved files.
Debug [LibraryScanner 1]: Committing transaction on "LIBRARY_SCANNER" result: true
Debug [LibraryScanner 1]: Detecting cover art for unscanned files.
Debug [LibraryScanner 1]: Scan finished cleanly
Debug [LibraryScanner 1]: Scan took: 16456057 ns. 14 unchanged directories. 0 changed/added directories. 0 tracks verified from changed/added directories. 0 new tracks.
Debug [Main]: BaseTrackCache(0x7fbbaea45b20) updateIndexWithQuery took 2 ms
Debug [Main]: LibraryTableModel(0x7fbbae77fc30) select() took 4 ms 281
||PaMacCore (vv)|| AudioIOProc()
Debug []: timeSinceLastCbSecs 1.38829e+06 callbackEntrytoDacSecs 0.013509 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.00145125 diff 1.38829e+06
Debug []: SSE: Denormals to zero mode already enabled
Debug []: SSE: Flush to zero mode already enabled
Debug []: Denormals to zero mode is working
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00140868 callbackEntrytoDacSecs 0.013538 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.013509 diff -1.36413e-05
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00156359 callbackEntrytoDacSecs 0.0134267 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.013538 diff 1.10082e-06
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.0014545 callbackEntrytoDacSecs 0.0134235 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134267 diff 3.28285e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00146521 callbackEntrytoDacSecs 0.0134095 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134235 diff -5.61314e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00143655 callbackEntrytoDacSecs 0.0134242 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134095 diff 4.89305e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00145196 callbackEntrytoDacSecs 0.0134235 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134242 diff -2.82923e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.0014224 callbackEntrytoDacSecs 0.013453 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134235 diff 6.47889e-07
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00146689 callbackEntrytoDacSecs 0.0134368 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.013453 diff -5.37307e-07
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00146186 callbackEntrytoDacSecs 0.0134261 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134368 diff -9.6814e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00145875 callbackEntrytoDacSecs 0.0134186 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134261 diff 1.77263e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00141868 callbackEntrytoDacSecs 0.0134512 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134186 diff 2.08382e-08
||PaMacCore (vv)|| AudioIOProc()
Debug [Engine]: timeSinceLastCbSecs 0.00149518 callbackEntrytoDacSecs 0.0134073 bufferSizeSec 0.00145125 m_lastCallbackEntrytoDacSecs 0.0134512 diff 6.37054e-08
||PaMacCore (vv)|| AudioIOProc()
…